requirements.txt
requests==2.24.0
requests-toolbelt==0.9.1
python-dotenv==0.20.0
Flask==1.1.2
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
itsdangerous==1.1.0